+++ src/Commands/Pocket/PocketInit.ts
@@ -21,7 +21,7 @@ export default class extends Command {
 
     async init(message: Message) {
         if(!super.hasPermissions(message)) {
-            return message.channel.send(this.Embed.missing_perms.call(this));
+            return message.channel.send(this.Embed.missing_perms());
         } 
 
         const client = await pool.pocket.connect();